refactor: split internal/controllers/interfaces.go grab-bag [sctl-b9i]
interfaces.go now holds only SolarController. ModbusClient and
MetricsCollector move into the epever package next to their only
consumers, with SetMetrics taking *ControllerStatus instead of
interface{}. The concrete RTU client is renamed SerialModbusClient so
the interface keeps the ModbusClient name. The epever-specific mocks
move into epever test files since the shared mocks package can no
longer name those interfaces without an import cycle.
