+++ internal/controllers/interfaces.go
@@ -1,8 +1,6 @@
 package controllers
 
 import (
-	"context"
-
 	"github.com/gin-gonic/gin"
 )
 
@@ -17,46 +15,4 @@ type SolarController interface {
 
 	// Close performs cleanup and releases resources held by the controller.
 	Close() error
-}
-
-// ModbusClient defines the interface for Modbus communication operations.
-// This abstraction allows for testing without physical hardware.
-type ModbusClient interface {
-	// ReadInputRegisters reads input registers from the Modbus device.
-	ReadInputRegisters(ctx context.Context, address, quantity uint16) ([]byte, error)
-
-	// ReadHoldingRegisters reads holding registers from the Modbus device.
-	ReadHoldingRegisters(ctx context.Context, address, quantity uint16) ([]byte, error)
-
-	// WriteSingleRegister writes a single holding register to the Modbus device.
-	WriteSingleRegister(ctx context.Context, address, value uint16) ([]byte, error)
-
-	// WriteMultipleRegisters writes multiple holding registers to the Modbus device.
-	WriteMultipleRegisters(ctx context.Context, address, quantity uint16, value []byte) ([]byte, error)
-
-	// ReadCoils reads coil status from the Modbus device.
-	ReadCoils(ctx context.Context, address, quantity uint16) ([]byte, error)
-
-	// WriteSingleCoil writes a single coil to the Modbus device.
-	WriteSingleCoil(ctx context.Context, address, value uint16) ([]byte, error)
-
-	// Close closes the Modbus client connection.
-	Close()
-}
-
-// MetricsCollector defines the interface for collecting and exposing metrics.
-// This abstraction allows for testing without the Prometheus global registry.
-type MetricsCollector interface {
-	// IncrementFailures increments the failure counter.
-	IncrementFailures()
-
-	// IncrementWriteFailures increments the write failure counter (if supported).
-	IncrementWriteFailures()
-
-	// IncrementRegisterFailure increments the register-specific failure counter.
-	IncrementRegisterFailure(address uint16, registerType string)
-
-	// SetMetrics updates all metrics based on the provided status.
-	// The status parameter should be a pointer to a controller-specific status struct.
-	SetMetrics(status interface{})
 }
